Start moving git diff management into git store
