+++ crates/remote_server/src/headless_project.rs
@@ -89,14 +89,15 @@ impl HeadlessProject {
 
         let environment = project::ProjectEnvironment::new(&worktree_store, None, cx);
         let git_store = cx.new(|cx| {
-            GitStore::local(
+            let mut store = GitStore::local(
                 &worktree_store,
                 buffer_store.clone(),
                 environment.clone(),
                 fs.clone(),
-                session.clone().into(),
                 cx,
-            )
+            );
+            store.shared(SSH_PROJECT_ID, session.clone().into(), cx);
+            store
         });
         let prettier_store = cx.new(|cx| {
             PrettierStore::new(
